FilePath(File()) does not respect the OS of the remote machine.
File() is always run on the local machine (master) so any paths run through it would be converted to the OS of the master node.  This was problematic when the remote and the master had different OS.   Using FilePath(File) also converted based upon the OS of the master.   This change uses FilePath(Channel, String) to use the channel to the remote server so that the path is correctly configured for the remote OS, irrespective of the master OS.