Fix update command changing workspace configuration
Recent versions of the plastic CLI switch the workspace selector
to point to the latest change in branch by default when calling "cm
update". This breaks the ability to sync to a given changeset and causes
more comples builds to potentially use different versions.

This restores the old behavor of "cm update" by adding the somewhat
undocumented flag "--changeset" without value.