Fix file not found issues
We need to use FilePath instead of File. FilePath represents files
in a local or remote machine. It prevents the kind of issues we're
experiencing since it provides the tools to operate on remote files.
The code is generally run on the 'master' machine but remotely on
the agents. The filesystem context is always the one from the
'master' machine.

Signed-off-by: Miguel González <mgonzalez@codicesoftware.com>