Add auto-provisioning and credentials capabilities
The `cm` tool is now managed as a Jenkins tool. This enables the plugin
to install it automatically, downloading the client bundle from
the website plasticscm.com.

This also enables users to specify credentials per project instead of
relying on the machine configuration. Credentials are retrieved using
the Jenkins Credentials API. This feature requires running client
version 10.0.16.6112 or newer to work.

Signed-off-by: Miguel González <miguel.gonzalez@unity3d.com>