Change order of priority
`shelveset` or `label` should be the taken into consideration first.
Then, if none of them is set, we'd use the branch-type selector. The
plugin will include the changeset number if set.

This order makes sense because we can't decouple changeset from branch.
As  we're relying on selectors (as opposed to specs), the selector needs
to know about the branch that contains the changeset.

Signed-off-by: Miguel González <miguel.gonzalez@unity3d.com>