Replace old @RequirePOST with @POST
According to
https://www.jenkins.io/doc/developer/security/form-validation/#protecting-from-csrf,
the @RequirePOST is an older annotation for preventing CSRF which
presents a form for the user to resubmit the request using POST.

However, in this case we can safely use the newer @POST annotation,
which limits processing to the POST verb. Other verbs will return a 404
response. It's the recommended action for validation methods.

Signed-off-by: Miguel González <miguel.gonzalez@unity3d.com>